integration-dispatcher/pyproject.toml
[project]
name = "self-service-agent-integration-dispatcher"
version = "0.1.0"
description = "Multi-tenant Integration Dispatcher for the self-service agent quickstart"
authors = [
    {name = "Self-Service Agent Team", email = "team@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "cloudevents>=1.10.0",
    "httpx>=0.25.0",
    "structlog>=23.2.0",
    "python-multipart>=0.0.22",
    "jinja2>=3.1.0", # For email templates
    "aiosmtplib>=3.0.0", # For email sending
    "aioimaplib>=1.0.0", # For email receiving (IMAP)
    "slack-sdk>=3.26.0", # For Slack integration
    "aiohttp>=3.13.3", # Required by slack-sdk for async operations
    "orjson>=3.11.7",
    "langchain-core>=1.2.11",
    "self-service-agent-shared-models",
    "self-service-agent-shared-clients",
    "protobuf>=6.33.5",
    "opentelemetry-exporter-otlp-proto-http>=1.37.0",
    "opentelemetry-instrumentation-httpx>=0.58b0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "tracing-config",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/integration_dispatcher"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"

# Configure mypy to use source directories for local packages
mypy_path = "../shared-models/src:../shared-clients/src:../tracing-config/src"

[[tool.mypy.overrides]]
module = [
    "opentelemetry.*",
]
follow_imports = "skip"

[[tool.mypy.overrides]]
module = [
    "aioimaplib.*",
]
ignore_missing_imports = true

[tool.uv.sources]
self-service-agent-shared-models = { path = "../shared-models" }
self-service-agent-shared-clients = { path = "../shared-clients" }
tracing-config = { path = "../tracing-config" }

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.2.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=5.13.0",
    "mypy>=1.17.1",
]
